Admin Store FPS bandaid
Admin store for now only has 1x and 3xT weapons, and base and lostech variants. This makes it actually usable and browsable for testing if generation works properly, even if some stuff is unavailable for testing for now.

I'll disable this behavior again at some point, or expose it to the user.