Bounty tokens/pilots price changes
The first of sweeping changes to value of bounties and pilots, meant to fit the gameplay and narrative better in tandem with spreadsheet changes to be committed right after this. The $1 prices are intentional, ask me about it if needed.
Also, commented out my dev commodities from being sold on Denver.