Swap npc guns to manual (#80)
* Properly clean out empty/unfinished cols from shiparch

* Untangle PC and NPC guns 1/2

NPC guns will probably get a one-by-one handcraft-y treatment, or so the last discussion sounded

* Create sort_ini

This is for more human-like sorting in the generated .inis with more than one [block] type.

* Disallow Comments from blocks

I'll change it later so they go outside the block or something. For now, they just get cleaned out

* Revert "Untangle PC and NPC guns 1/2"

This reverts commit b2c6ed54e565457efb6f790d423b241bf45f11f3.

* Do the sorting after generation

also: fixed the template to make the sorting more stable. probably definitely works.

* Separate commit for sorting-related INI changes

* Prettier numbers for generated sheets

* More pretty numbers

* Apparently this fixes a bug where some infocards are missing

?????????????????????????????????????

* Comments are back

and on top of blocks now!

Now I will add Comment columns to *all* sheets, front and back, and rewrite the human-like sorting so it starts blocks at the comments and not the square brackets. For now, human-like sorting is temporarily disabled again.

* Comments are sorted properly

* Untangle PC and NPC guns (2/2)

I did some ugly things but this should be done now.

* aux scaling rules removed

* Resulting INI changes

* Revert "Resulting INI changes"

This reverts commit fa9f4b4da83606dbe8829e31e73dd30b097eca1f.

* Stop looking for Aux Scaling Rules

* Unfuck nan comments

* no nan lines in blocks even if they somehow snuck past

---------

Co-authored-by: Rupert Walker <RupertCW@gmail.com>