New lootbox item
Created new lootbox item to be used as a loottables.dll trigger. However, through setting this up I've noticed this and any other new commodity made via the spreadsheet seems to crash the game when a rookie carrying the cargo gets destroyed.
Not sure why this is, need help running it down.