Continuation of prev commit on new Mule loadouts
Rest of the entries for making the new Mule loadouts actually ready to be spawned. This should've probably been bundled in the prev commit, oops.

While I'm at it with mistakes, forgot to note I also changed how many pilots spawn in Xenos Independences - the intention is to show that those frigates aren't run by a single pilot but by a sizeable crew, which is why each has around 15 souls aboard right now. We can replace the pilots item with an actual ship crew item later for polish.