Buffed Navy Scouts 2 encounter
Made this encounter into a 4-ship wing, with 2 vets and 2 rookies, using the new tactical wedge formation. They currently are set to appear in the Big Yen subzone.
Barred them from showing up around Denver as they were previously doing; the tactical wedge formation is suitable for combat, but not for passive duty near the planet as the wideness of the formation will inevitably drag members into the planet's deathzone.