Major Silverton rework
Rebuild of Silverton field's encounters from the ground up. The field has a diverse variety of encounters available, and balancing the weight of each of their occurence rates will be an ongoing process.

Added new encounter files for main mobs, reinforcement waves and trickling in adds for both Xenos and Navy.
Adjusted li03.ini to balance their spawn rates and specific exclusion caps.
Also snuck in a small change to how the formatting is done for Cheyenne's scouts_3 reinforcements.