Committed a bunch of stuff missing from repo?
I might just be stupid? But I tried to play on dev-live on our main server and some shit instantly broke compared to my localhost, when I compared the repo to my install I noticed a lot of inis and such in the BMOD folder that weren't on the repo?

Revert this and let me know if it was intentional but I can't see how or why it would be