"Tikas" Supply Base added
Introduced a small (un)dockable asteroid base in Alamosa next to the Ludlow frigate and the jumphole, named "Tikas". Gave it a WIP loadout of a few PD Blasters and an SRM turret.
Will tweak its firepower more suitably in the future, including giving the base better firing angles and hardpoint positions.