Increased AC/Missile's damage  typesagainst shields
Change to an absolutely ancient adjustment I made in BMOD's first versions that made the damage types for autocannons (BMOD_KINETIC_01) and missiles (BMOD_EXPLOSIVE_01) do fractions of their overall damage to shields.
This was necessary back in the day when we didn't yet have seperate damage values vs. hull and energy, but as far as I can tell I just forgot about this and never reverted it after we got that functionality. Considering we can already set shield damage to be whatever we want now, it's long past time to set this back so that these weapons do their full listed anti shield damage, however slight that damage may be.