Renamed _xdm blasters to _xp1
This is a refactor to change the name of the existing _xdm blasters which are used on every elite fighter + are for sale on Pueblo station. Renaming these to match the intended naming scheme now before this becomes a bigger problem in the future.

Keep your eyes out for any problems from this but I tried to be thorough in covering every base. Also adapted lootprops in this commit to account both for this and the previous _xp2 and _xp3 changes.