Missile variant reformat
Again, another renaming refactor, this part is for missiles. These were only called on the dev station and on the unused ace loadouts and so shouldn't even be in the game right now anyway.
Amusingly, this one took the longest because I didn't realise the Rheinland fighter on the main menu cutscene had a _plu LRM, which took a bit of debugging to figure out.