Gear renaming
This should be the last in the series - [+] Variants of CM Dropper, Cruise Disruptor and Repair Mine dealt with. There should no longer be any old variant gear in the mod that players could or can get a hold of. The only thing left with _plu should be a one-off placeholder turret solars are using, and as players can't get that, it should be able to be fixed later.