Hotfix for 0.0.13 (#108)
* fix: Adjust the Rhino's cruise start sound to match it's charge time

* Blaster variant changes

Removed unused placeholder variants: xrr, xee, xmv, xer and xlt
Added new plus variants: xp2 and xp3 ([++] and [+++])
Removed a commented out line about xmv in loadouts_npc and changed a solar that was using xmv's

Set Civilian xp2 and xp3 equipment as for sale on the Dev Station

Will change _xdm to be named as _xp1 shortly for consistency's sake

* Renamed _xdm blasters to _xp1

This is a refactor to change the name of the existing _xdm blasters which are used on every elite fighter + are for sale on Pueblo station. Renaming these to match the intended naming scheme now before this becomes a bigger problem in the future.

Keep your eyes out for any problems from this but I tried to be thorough in covering every base. Also adapted lootprops in this commit to account both for this and the previous _xp2 and _xp3 changes.

* Shield variant renaming

Same as last commits but this time I'm renaming shield variants from the _plu suffix into _xp1 instead. These shields are in use on elite NPCs and for sale on Pueblo.

* Missile variant reformat

Again, another renaming refactor, this part is for missiles. These were only called on the dev station and on the unused ace loadouts and so shouldn't even be in the game right now anyway.
Amusingly, this one took the longest because I didn't realise the Rheinland fighter on the main menu cutscene had a _plu LRM, which took a bit of debugging to figure out.

* Aux weapon renaming

Another one. This part is renaming all the Aux weapons. Only one NPC loadout used this, and that's the unused ace who isn't in the game yet; these weapons weren't available in any normal way.

* Gear renaming

This should be the last in the series - [+] Variants of CM Dropper, Cruise Disruptor and Repair Mine dealt with. There should no longer be any old variant gear in the mod that players could or can get a hold of. The only thing left with _plu should be a one-off placeholder turret solars are using, and as players can't get that, it should be able to be fixed later.

* Implemented desired values for [+] Gears

[+] versions of CM dropper, CD and Repair Mine tweaked.
+ CM Dropper gets double ammo and +5% flare effectiveness.
+ CD gets -1.5s refire rate.
+ Repair Mine gets -3s refire rate and -3s mine detonation wait time.

* Fix two idiotic bugs regarding aux generation

* Plus Gear further progress

- Set [+] versions of CM Droppers, CDs, Repair Mines and LRMs/SRMs/their Pylons all up for sale at intended high prices on Pueblo. Copied this generic [+] gear to be for sale on Ouray as well.
- Made [+] blasters of other faction lines available on appropriate bases
- Barred the RAC-10 from sale, it wasn't intended to be in for playtesters
- Tweaked the stats of the [+] gear

* glitchy version of ammo-on-demand auxs

* maybe fixed bad generation

* Revert "maybe fixed bad generation"

This reverts commit 39b2a127ff8f76640c76001d295a792f08d37896.

* Revert "Merge branch 'dev/0.0.13' of https://github.com/better-modernized-combat/bmod-client into dev/0.0.13"

This reverts commit cb427401d026b73faad90b7008de01a371ddf96d, reversing
changes made to 44018e11ea7faf22df82f84e4148119990295f2c.

* Revert "glitchy version of ammo-on-demand auxs"

This reverts commit 44018e11ea7faf22df82f84e4148119990295f2c.

* Recommit of [+] gear progress, and a fix

Recommit of prior commit that got reverted, also, new faction blaster [+] variants didn't have good prices, fixed that to be appropriate.

* Committing new sheet automation for Aux Guns

Thank you Katz for new PC Aux Gun sheet functionality, committing what the sheet spits now

* Revert "Committing new sheet automation for Aux Guns"

This reverts commit 8f2270a2f936127abfb9fcfde5aae9b52e96f43d.

* Made Mining Ship and Roid Base solars indestructible

- death fuse commented out on miner and roid base solararches, replaced with a drop loot surprise fuse at 10k HP
- Added a lot of bounty vouchers that drop from Ludlow and Tikas when you get them below that 10k HP number

* Stack size for bounties increased

Up from 1 to 50, so that they drop in bunches properly from the new Ludlow/Tikas loot

* 0.0.13 Changelog, ready to go

Build's done!

---------

Co-authored-by: Beaglerush <beaglerush@gmail.com>
Co-authored-by: FJonske <frederic.jonske@rwth-aachen.de>