Rep tweaked for short-term Xenos swap method
New MP characters will start with -1 Xenos rep and 0.1 Navy rep, so that if they blow up a bunch of storage depots, they'll become hostile with Navy first, then tolerated neutral with Xenos second.
The result is a short-term way for new characters to easily swap to Xenos and play as pirates. We'll make this more ideal in the future, of course.