Replaced Trent with anonymous mute for new MP characters
Replaced default Trent MP character with a helmeted, visor'd, silent person instead.
The goal is to change our default MP char so they're a better representative of every player instead of just white blonde dudes. Would be good to get custom choices and new voicework done in the far future but for now this is enough to improve things I think.