Added missing basic weapons to Denver sale
Denver did not have PPCs, L Lasers or Twin SRM/LRM launchers. Players do not currently have any intuitive way to know they should go looking on other bases to buy these guns, and are likely to just believe if it's not on Denver it's not out there to be found.
So for now until we have proper direction and signposting, making sure Denver is a one stop shop for the basics.