0.0.17 hotfix (#188)
* feat: Migrate out-of-use systems to LancerEdit's file standard.
* Adjustments to Xenos HF pilot behaviour

Existing bmod_pilots have proven to be too timid and scared of being pursued lately, dodging too often. I don't know when this behaviour became a thing but I've improved upon it today:

- There are now three types of pilot personality: Default, Aggro and Coward
- Coward is the old behaviour we've been using where they Evade a lot if you get behind them
- Aggro pilots almost never get into Evade behaviour
- Default pilots are a balance in between

For now this is only applied to Xenos Falcon and Angler d1-6 enemies. I will continue to add this to all the ships over time, including new pilot behaviours for the LFs and VHFs finally

* maybe hotfix for missing npc +ammo lootprops entries
* chore: Update Changelog

Co-authored-by: Beaglerush <beaglerush@gmail.com>
Co-authored-by: FJonske <frederic.jonske@rwth-aachen.de>