Added new Mobile Repair Bay as a Solar
In game now as bm_navy_mrb with a test placement of it south of Maverick station. Has damage fuses but not a death one so it shouldn't fully die. Has a name ID but no infocard yet.
Also adjusted Lazrius frigate's name ID and added a new one for Solar versions.