Tweaked MRB hardpoints and added loadout
Added a loadout for the new Mobile Repair Base with two energy PDTs and a specific setup for its many lights. Tweaked the MRB model's hardpoints; re-made the PDT HPs as they were fixed instead of revolute, removed the third one and shifted the position of the second one to better cover the model's rear.
Pitched the bay lights to be visible from the direction you look at them. Added new lights to the ship's front, rear and sides.

Note that the rear PDT HP has obstructions on its left and right and could use a taller nub mount so it can shoot beyond them.