CSV M Energies upgraded to L
CSV was too undergunned for a player to have a reason to choose it, even with its support capability. Expanding its two blaster HPs from M to L Energy brings its basic firepower more in line with a comparable HF, while still limiting it from having the same level of special weaponry in missile and ballistic.
It also allows potentially ill-advised but interesting builds like running PPCs and L Lasers, which seems fitting for the different kind of role it can fill out as a support fighter with repair guns.