CSVs added to Xenos NPC lineup
Rookie/Pilot/Vet/Elite versions of CSV support fighters added to the Xenos. They can show up in any encounter tagged for the generic bm_class_fighter, or specifically be called as bm_class_support_fighter in the same way light or heavies can be specified.

I'm going to use these CSVs to fill out encounters as "three quarters of an enemy" where I want more numbers but don't want to go all the way to adding a full combat-dedicated fighter to the odds against the players. Often as a rookie.