Trying new experimental "Ship Variant" approach
Cloned copies of existing ships, in order to make it more obvious to the player what loadout and veterancy the enemy they're facing is bringing.
This makes available two Falcon variants intended for Xenos Elites, a Close Quarters and a Long Range version, with their own specific loadouts and IDs.